requirements.txt
python-dotenv
black==24.10.0
gymnasium==0.29.1
tensorboard==2.18.0
tqdm==4.66.5
pyyaml==6.0.2
wandb==0.18.5
joblib==1.4.2
torchsummary==1.5.1
stable-baselines3==2.4.1
prettytable==3.13.0
seaborn==0.13.2
ConfigSpace==1.2.1